+++ src/store/actions/title.js
@@ -1,5 +1,10 @@
 // @flow
-import { TITLE_SET, TITLE_RESET } from '../constants'
+import { APP_NAME_SET, TITLE_SET, TITLE_RESET } from '../constants'
+
+export type SetAppNameAction = {
+  type: 'APP_NAME_SET',
+  payload: string
+}
 
 export type SetTitleAction = {
   type: 'TITLE_SET',
@@ -13,6 +18,11 @@ export type ResetTitleAction = {
 export const setTitle = ({ title, propsList }: { title: string, propsList: AppTitleProps[] }): SetTitleAction => ({
   type: TITLE_SET,
   payload: { title, propsList }
+})
+
+export const setAppName = (name: string): SetAppNameAction => ({
+  type: APP_NAME_SET,
+  payload: name
 })
 
 export const resetTitle = (): ResetTitleAction => ({ type: TITLE_RESET })