+++ src/components/Routes.js
@@ -1,0 +1,44 @@
+// @flow
+
+import React from 'react';
+import { Route, Switch } from 'react-router-dom'
+import { connect } from 'react-redux'
+import NoComponent from './NoComponent'
+import coreInstance from '../coreInstance'
+import type { State } from '../store'
+import { selectRouteIsAllowed } from '../store/selectors'
+
+type RoutesProps = {
+  routeIsAllowed: boolean,
+}
+
+const mapRoutesModule = () => {
+  const modules = coreInstance.getModules()
+  return modules.map(module => (
+    <Route key={module.namespace} path={'/' + module.namespace} component={module.RootComponent} />
+  ))
+}
+
+class Routes extends React.Component<RoutesProps> {
+  componentDidMount() {
+    coreInstance.subscribe('registerModule', () => {
+      this.forceUpdate()
+    })
+  }
+
+  render() {
+    const { routeIsAllowed } = this.props;
+    return (
+      <Switch>
+        {routeIsAllowed && mapRoutesModule()}
+        <Route path={'/'} component={NoComponent} />
+      </Switch>
+    )
+  }
+}
+
+export default connect((state: State) => {
+  return {
+    routeIsAllowed: selectRouteIsAllowed(state)
+  }
+})(Routes)