+++ src/store/index.js
@@ -4,14 +4,15 @@ import menuReducer, { type MenuState } from './reducers/menu'
 import notificationsReducer, { type NotificationState } from './reducers/notifications'
 import uiReducer from './reducers/ui'
 import { changeTitleMiddleware } from './middlewares/title'
+import pageFilterMiddleware from './middlewares/pageFilter'
 import { setAppName } from './actions/title'
 import appTitleReducer, { type AppTitleState } from './reducers/title'
 import thunk from 'redux-thunk'
-import { connectRouter, routerMiddleware } from 'connected-react-router'
+import { connectRouter, routerMiddleware, RouterState } from 'connected-react-router'
 import history, { APP_PATH_PREFIX } from './history'
-import CoreInstance from '../coreInstance'
+import coreInstance from '../coreInstance'
 import * as constants from './constants'
-import { type FSA } from '../core'
+import Core, { type FSA } from '../core'
 import {
   notificationStorageKey,
   saveNotificationsMiddleware,
@@ -19,63 +20,72 @@ import {
   unzipData
 } from './middlewares/notifications'
 import type { UiState } from './reducers/ui'
+import type { PageFilterState } from './reducers/pageFilter'
+import pageFilterReducer from './reducers/pageFilter'
 
 export type State = {
   menu: MenuState,
   appTitle: AppTitleState,
   notifications: NotificationState,
-  ui: UiState
+  ui: UiState,
+  pageFilter: PageFilterState,
+  router: RouterState
 }
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
 
-const store = createStore(
-  connectRouter(history)(
-    combineReducers({
-      menu: menuReducer.reduce,
-      appTitle: appTitleReducer,
-      ui: uiReducer,
-      notifications: notificationsReducer
-    })
-  ),
-  {
-    notifications: unzipData(localStorage.getItem(notificationStorageKey))
-  },
-  composeEnhancers(
-    applyMiddleware(
-      thunk,
-      routerMiddleware(history),
-      menuReducer.middleware,
-      saveNotificationsMiddleware,
-      changeTitleMiddleware
+export const createCoreStore = (core: Core) => {
+  const store = createStore(
+    connectRouter(history)(
+      combineReducers({
+        menu: menuReducer.reduce,
+        appTitle: appTitleReducer,
+        ui: uiReducer,
+        notifications: notificationsReducer,
+        pageFilter: pageFilterReducer
+      })
+    ),
+    {
+      notifications: unzipData(localStorage.getItem(notificationStorageKey))
+    },
+    composeEnhancers(
+      applyMiddleware(
+        thunk,
+        routerMiddleware(history),
+        menuReducer.middleware,
+        saveNotificationsMiddleware,
+        changeTitleMiddleware,
+        pageFilterMiddleware
+      )
     )
   )
-)
 
-CoreInstance.subscribe('registerModule', () => {
-  const modules = CoreInstance.getModules()
-  for (const module of modules) {
-    const added = menuReducer.processModule(module)
-    if (added) {
-      store.dispatch({
-        type: constants.RESET,
-        payload: {
-          namespace: module.namespace,
-          path: window.location.pathname.slice(APP_PATH_PREFIX.length)
-        }
-      })
+  core.subscribe('registerModule', modules => {
+    for (const module of modules) {
+      const added = menuReducer.processModule(module)
+      if (added) {
+        store.dispatch({
+          type: constants.RESET,
+          payload: {
+            namespace: module.namespace,
+            path: window.location.pathname.slice(APP_PATH_PREFIX.length)
+          }
+        })
+      }
     }
-  }
-})
+  })
 
-CoreInstance.subscribe('setAppName', (name: string) => {
-  if (typeof name === 'string' && name) {
-    store.dispatch(setAppName(name))
-  }
-})
+  core.subscribe('setAppName', (name: string) => {
+    if (typeof name === 'string' && name) {
+      store.dispatch(setAppName(name))
+    }
+  })
 
-CoreInstance.subscribe('dispatchToken', (token: FSA) => store.dispatch(token))
+  core.subscribe('dispatchToken', (token: FSA) => store.dispatch(token))
 
-startTimerCheckNotifications(store)
+  startTimerCheckNotifications(store)
 
-export default store
+  return store
+}
+
+export default createCoreStore(coreInstance)