+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 - Add cache-control http header for /static route
+- Fixed state where could be more than one selected menu items.
 
 ## [6.4.0] - 2020-03-17